doesn't force users to refresh if last.fm doesn't work for darksabrelord

+++ static/script.js
@@ -31,9 +31,11 @@ var CACHED_USERS = JSON.parse(window.localStorage.cached_users || "{}");
 
                         // TODO: Show erorr message ;)
                     } else {
-                        alert("Last.fm took too long to respond.\n\nLet's refresh the page and try again!");
-                        window.localStorage.clear();
-                        window.location.reload();
+                        var refresh = confirm("Last.fm took too long to respond.\n\nPress OK to refresh the page and try again, or Cancel to use the page as it is.");
+                        if (refresh) {
+                            window.localStorage.clear();
+                            window.location.reload();
+                        }
                     }
                     return;
                 }
